requirements.txt
requests
beautifulsoup4
SQLAlchemy
psycopg2-binary
python-dotenv
